finish renaming integrations folder stuff, and fix lint
