+++ src/mods/hooks/useRequestPermissionsAction.ts
@@ -17,10 +17,10 @@
 
 import { type ModViewItem } from "@/types/modTypes";
 import { useCallback } from "react";
-import { type OptionsState } from "@/store/extensionsTypes";
 import { selectExtensionsFromMod } from "@/utils/modUtils";
 import { useSelector } from "react-redux";
 import useModPermissions from "@/mods/hooks/useModPermissions";
+import { type ModComponentState } from "@/store/extensionsTypes";
 
 function useRequestPermissionsAction(
   modViewItem: ModViewItem
@@ -30,7 +30,8 @@ function useRequestPermissionsAction(
   // Without memoization, the selector reference changes on every render, which causes useModPermissions
   // to recompute, spamming the background worker with service locator requests
   const memoizedExtensionsSelector = useCallback(
-    (state: { options: OptionsState }) => selectExtensionsFromMod(state, mod),
+    (state: { options: ModComponentState }) =>
+      selectExtensionsFromMod(state, mod),
     [mod]
   );
