+++ src/pageEditor/fields/FormRendererOptions.tsx
@@ -32,7 +32,7 @@ import {
 } from "@/bricks/renderers/customForm";
 import AppApiIntegrationDependencyField from "@/components/fields/schemaFields/AppApiIntegrationDependencyField";
 import { type ModComponentFormState } from "@/pageEditor/starterBricks/formStateTypes";
-import { produceExcludeUnusedDependencies } from "@/components/fields/schemaFields/integrations/serviceFieldUtils";
+import { produceExcludeUnusedDependencies } from "@/components/fields/schemaFields/integrations/integrationDependencyFieldUtils";
 import FieldTemplate from "@/components/form/FieldTemplate";
 import Select, { type Options } from "react-select";
 import FORM_FIELD_TYPE_OPTIONS from "@/pageEditor/fields/formFieldTypeOptions";
